iac-terraform/terraform-docker-flask/flask_app/app/requirements.txt
#flask
#psycopg2-binary
Flask==2.3.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0

